Remove CMake and instead perform all logic in utils/build
